pyproject.toml
[project]
name = "streamduck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.3.2",
    "streamlit>=1.49.1",
    "python-dotenv>=1.0.0",
]
